Revert fp16 -> fp32 weights (fixes 'Attempting to unscale FP16 gradients'); lower batch_size 8->4 for OOM margin

+++ training/finetune_hf.py
@@ -103,7 +103,7 @@ from transformers import AutoModelForCausalLM, AutoTokenizer
 
 print(f"Loading {MODEL_NAME} ...")
 model = AutoModelForCausalLM.from_pretrained(
-    MODEL_NAME, trust_remote_code=TRUST_REMOTE, torch_dtype=th.float16
+    MODEL_NAME, trust_remote_code=TRUST_REMOTE, torch_dtype=th.float32
 )
 tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME, trust_remote_code=TRUST_REMOTE)
 if tokenizer.pad_token is None:
